Fix all clang compiler warnings in demos and config/util
Fixed multiple categories of clang warnings across demo programs and utilities:

1. Void-pointer-to-enum-cast warnings:
   - demos/lib/Exm/String.c: Use XmNO_TEXT instead of (XmTextType)NULL

2. Int-to-void-pointer and void-pointer-to-int cast warnings:
   - Added intermediate (long) casts for proper type conversion
   - Fixed warnings in 25+ demo programs including:
     * ButtonBox/bboxdemo.c (5 warnings)
     * FontSel/callbacks-c.c
     * TabStack/callbacks.c, bxutils.c (4 warnings)
     * Tree/bxutil-c.c, tree.c (5 warnings)
     * ColorSel/colordemo.c, util-c.c (4 warnings)
     * Ext18List/callbacks-c.c, util-c.c (5 warnings)
     * IconB/iconbuttondemo.c, misc.c (4 warnings)
     * Paned/paned.c (2 warnings)
     * Unsupported programs: xmforc.c (3), motifshell.c (7)
     * Program guide demos: interface.c files (6 warnings)

3. Function pointer type errors:
   - config/util/makestrs.c: Fixed incompatible function pointer types
     * headerproc[]: void (*)() -> void (*)(FILE*, File*)
     * sourceproc[]: void (*)() -> void (*)(int)

All warnings were fixed using proper type casting with intermediate
(long) conversions to ensure compatibility across different architectures
and maintain code safety. The fixes preserve the original functionality
while eliminating compiler warnings.

+++ demos/doc/programGuide/ch08/Notebook/interface.c
@@ -96,7 +96,7 @@ HelpCB(Widget   w,
        XtPointer cb
       )
 {
- int       what_kind_of_help = (int)cd;
+ int       what_kind_of_help = (int)(long)cd;
  char      help_string[400];
  XmString  hs_as_cs;
  Widget    dialog_general_help;
