+++ demos/doc/programGuide/ch17/simple_drag/simple_drag.c
@@ -272,7 +272,7 @@ HelpCB(Widget   w,
        XtPointer cb
       )
 {
- int       what_kind_of_help = (int)cd;
+ int       what_kind_of_help = (int)(long)cd;
  char      help_string[500];
  XmString  hs_as_cs;
  Widget    dialog_general_help;
