+++ demos/programs/FontSel/callbacks-c.c
@@ -70,7 +70,7 @@ XtPointer call;
     Cardinal argcnt;
     XmString xmstring=NULL;
     static Widget info = NULL;
-    int explain = (int) client;
+    int explain = (int)(long) client;
 
     if (info == NULL) {
 	Widget temp;
